Split long lines into multi-lines texts and trim trailing spaces. Those splits do not make any difference in the HTML output when viewed from a browser. There is no change in the content (any edition proposals will be the subject of separated pull requests). The intend of this reformatting is to enable smaller differences in future pull request submissions. For example given that 'git diff' shows changes on a line-by-line basis, if we propose to change a single word in the middle of a paragraph, such small change will be easier to see if the modified line reported by Git is short.
This commit adds new-line characters at "natural" points as much as possible, for example at the end of a sentence or after a comma. This approach tends to produce smaller diff. For example modifying a sentence does not appear if we also modified the sentence that immediately follow it if those two sentences are on separated lines.